requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
docx2txt==0.8
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
nltk==3.9.1
numpy==2.1.3
pandas==2.2.3
pillow==11.0.0
pypdf==5.1.0
PyPDF2==3.0.1
PyPDF3==1.0.6
PyPDF4==1.27.0
python-dateutil==2.9.0.post0
pytz==2024.2
qrcode==8.0
regex==2024.11.6
six==1.16.0
tqdm==4.67.1
tzdata==2024.2
Werkzeug==3.1.3
